pyproject.toml
[project]
name = "music"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "slskd-api>=0.1.5",
    "spotdl>=4.2.11",
]
